load pretrained model from SimCLR
