+++ configs/semantic_kitti/imageBilinearPretrained.yaml
@@ -1,0 +1,47 @@
+# AUTO_RESUME: TRUE
+# RESUME_STATES: TRUE
+MODEL:
+  TYPE: "ImageSegBilinear"
+  DUAL_HEAD: False
+  NUM_CLASSES: 20 
+  late_feat_block_number: 11
+  USE_IMAGE: True
+  USE_LIDAR: False
+  USE_FUSION: False
+  IMAGE_PRETRAINED_PATH: "/home/user/logs/SimCLR/lightning_logs/version_54/checkpoints/epoch=15-step=8463.ckpt"
+DATASET:
+  TYPE: "SemanticKITTISCN"
+  TRAIN: ("train",)
+  VAL: ("val",)
+  TEST: ("test",)
+  SemanticKITTISCN:
+    preprocess_dir: "/home/user/SemanticKitti/preprocessed"
+    semantic_kitti_dir: "/home/user/SemanticKitti"
+DATALOADER:
+  NUM_WORKERS: 0
+OPTIMIZER:
+  TYPE: "Adam"
+  BASE_LR: 1e-4
+  WEIGHT_DECAY: 0.0005
+SCHEDULER:
+  TYPE: ""
+  # StepLR:
+  #   gamma: 0.1
+  #   step_size: 10
+  MAX_EPOCH: 100
+TRAIN:
+  BATCH_SIZE: 10
+  # SUMMARY_PERIOD: 2
+  # CHECKPOINT_PERIOD: 5
+  # CLASS_WEIGHTS: [1.89090012, 2.0585112, 3.1970535, 3.1111633, 1., 2.93751704, 1.92053733,
+  #                 1.47886874, 1.04654198, 1.78266561]
+  # CLASS_WEIGHTS: [0., 1.58003993, 3.69774469, 3.2460013,  2.65342029, 2.61079801, 3.27744058,
+  #                 3.48282471, 3.45874555, 1.,         2.07298878, 1.26831551, 2.65889542,
+  #                 1.37436805, 1.4891881,  1.03083152, 2.25629999, 1.51838281, 2.51986332,
+  #                 3.08564901]
+  FusionTransformer:
+    lambda_xm: 0.0
+VAL:
+  BATCH_SIZE: 10
+  # PERIOD: 5
+OUTPUT_DIR: "/home/user/logs/@"  #  @ will be replaced with config path, e.g. nuscenes/usa_singapore/FusionTransformer
